Implements UserDetails in User entity
