+++ src/main/java/ua/softserve/rv_028/issuecitymonitor/entity/User.java
@@ -3,11 +3,15 @@ package ua.softserve.rv_028.issuecitymonitor.entity;
 import org.hibernate.annotations.NaturalId;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import ua.softserve.rv_028.issuecitymonitor.dto.UserDto;
 import ua.softserve.rv_028.issuecitymonitor.entity.enums.UserRole;
 import ua.softserve.rv_028.issuecitymonitor.entity.enums.UserStatus;
 
 import javax.persistence.*;
+import java.util.Collection;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -16,7 +20,7 @@ import java.util.Set;
 @Table(name = "users")
 @SQLDelete(sql = "UPDATE users SET deleted = 'true' WHERE id = ?")
 @Where(clause = "deleted <> 'true'")
-public class User{
+public class User implements UserDetails{
 
 	@Id
 	@GeneratedValue
@@ -150,6 +154,7 @@ public class User{
 		this.lastName = lastName;
 	}
 
+	@Override
 	public String getPassword() {
 		return this.password;
 	}
@@ -158,12 +163,13 @@ public class User{
 		this.password = password;
 	}
 
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
+	@Override
 	public String getUsername() {
 		return this.username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
 	}
 
 	public String getPhone() {
@@ -222,9 +228,53 @@ public class User{
 		return isDeleted;
 	}
 
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+		return prepareCollection(this.userRole);
+	}
+
+	@Override
+	public boolean isAccountNonExpired() {
+		return true;
+	}
+
+	@Override
+	public boolean isAccountNonLocked() {
+		return true;
+	}
+
+	@Override
+	public boolean isCredentialsNonExpired() {
+		return true;
+	}
+
+	@Override
+	public boolean isEnabled() {
+		return checkUserStatus(this.getUserStatus());
+	}
+
 	@PreRemove
 	public void delete() {
 		this.isDeleted = true;
+	}
+
+	private static boolean checkUserStatus(UserStatus status) {
+		return status == UserStatus.ACTIVE || status == UserStatus.UNCONFIRMED;
+	}
+
+	private Collection<GrantedAuthority> prepareCollection(UserRole userRole) {
+		HashSet<GrantedAuthority> roles = new HashSet<>();
+		if (userRole == UserRole.USER) {
+			roles.add(new SimpleGrantedAuthority(UserRole.USER.name()));
+		} else if (userRole == UserRole.MODERATOR){
+			roles.add(new SimpleGrantedAuthority(UserRole.USER.name()));
+			roles.add(new SimpleGrantedAuthority(UserRole.MODERATOR.name()));
+		} else if (userRole == UserRole.ADMIN){
+			roles.add(new SimpleGrantedAuthority(UserRole.USER.name()));
+			roles.add(new SimpleGrantedAuthority(UserRole.MODERATOR.name()));
+			roles.add(new SimpleGrantedAuthority(UserRole.ADMIN.name()));
+		}
+		return roles;
 	}
 
 	@Override
